requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
alembic
altair
annotated-types
anyio
appdirs
asgiref
asttokens
attrs
auth0-python
babel
backoff
backrefs
bcrypt
beautifulsoup4
blinker
build
cachetools
certifi
cffi
cfgv
charset-normalizer
chonkie
chroma-hnswlib
chromadb
click
cohere
colorama
coloredlogs
crewai
crewai-tools
cryptography
dataclasses-json
decorator
defusedxml
deprecation
distlib
distro
docker
docstring_parser
dotenv
durationpy
embedchain
et_xmlfile
executing
fastapi
fastavro
filelock
firecrawl-py
flatbuffers
frozenlist
fsspec
ghp-import
gitdb
GitPython
google-auth
googleapis-common-protos
gptcache
grpcio
h11
h2
hf-xet
hpack
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperframe
identify
idna
importlib_metadata
importlib_resources
instructor
ipython
ipython_pygments_lexers
jedi
Jinja2
jiter
joblib
json5
json_repair
jsonpatch
jsonpickle
jsonpointer
jsonref
jsonschema
jsonschema-specifications
kubernetes
lancedb
langchain
langchain-cohere
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langsmith
litellm
magika
Mako
Markdown
markdown-it-py
markdownify
markitdown
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mem0ai
mergedeep
mkdocs
mkdocs-get-deps
mkdocs-material
mkdocs-material-extensions
mmh3
monotonic
mpmath
multidict
mypy_extensions
narwhals
nest-asyncio
networkx
nodeenv
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
paginate
pandas
parso
pathspec
pdfminer.six
pdfplumber
pexpect
pillow
platformdirs
portalocker
posthog
pre_commit
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyJWT
pymdown-extensions
pypdf
pypdfium2
PyPika
pyproject_hooks
pyright
pysbd
python-dateutil
python-dotenv
pytube
pytz
pyvis
PyYAML
pyyaml_env_tag
qdrant-client
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
schema
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
streamlit
sympy
tabulate
tenacity
threadpoolctl
tiktoken
tokenizers
toml
tomli
tomli_w
torch
tornado
tqdm
traitlets
transformers
typer
types-requests
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uv
uvicorn
uvloop
virtualenv
watchdog
watchfiles
wcwidth
websocket-client
websockets
wrapt
yarl
zipp
zstandard
